Refactor Floor Label drawing #2348
